ml_dir/pyproject.toml
[project]
name = "hackathon-bordeaux"
version = "0.1.0"
description = "Resili-Score — pipeline de données géospatiales Bordeaux Métropole"
requires-python = ">=3.11"
dependencies = [
    "geopandas",
    "shapely",
    "numpy",
    "pandas",
    "scikit-learn",
    "shap",
    "python-dotenv",
    "jupyter",
    "seaborn",
]

[project.optional-dependencies]
dxf = ["ezdxf"]
